move signature checks back to engine.py
